requirements.txt
certifi==2023.7.22
et-xmlfile==1.1.0
exceptiongroup==1.1.3
influxdb-client==1.37.0
influxdb3-python==0.2.1
iniconfig==2.0.0
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pluggy==1.2.0
pyarrow==13.0.0
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
reactivex==4.0.4
requests==2.31.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
